UPDATE: created routes for my models
NEXT: AUTHENTICATION
